replaced modal with format in pdf
